# server/swagger/spec.yaml
openapi: 3.0.0
info:
  description: Definition of the operations available for the ASquare Dashboard
  title: ASquare APIs
  version: 0.1.0
servers:
  - url: http://localhost:3000/api
paths:
...
  /auth/password/change:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordResponse"
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Bad request. Some of the inputs provided to the request aren't valid"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: User password have expired
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Internal server error. Additional details will be contained on the server logs."
      requestBody:
        content:
          application/json:
            schema:
              properties:
                password:
                  type: string
      operationId: changePassword
      summary: Change password
      description: Change user or admin password if they are logged in
      tags:
        - Authentication
components:
  schemas:
    ResetPasswordResponse:
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
        - description
...
    ErrorResponse:
      properties:
        detail:
          description: Optional text that describes the error in more detail
          type: string
        instance:
          description: Optional UID to correlate message to logs
          type: string
        status:
          description: repeat of returned HTTP Status Code
          type: number
        title:
          description: >-
            Human readable summary message that may be displayed back to the
            user.
          maxLength: 1024
          minLength: 1
          type: string
        type:
          description: URI for type of error
          minLength: 1
          type: string
      required:
        - status
        - type
        - title
      type: object
...